going to backport this into IParser branch
