+++ src/parser/parsers/index.ts
@@ -1,4 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
+import * as Naive from "./naive";
+
 export { RecursiveDescentParser } from "./recursiveDescentParser";
+export { Naive };
